requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.32.3
beautifulsoup4==4.12.3
numpy==2.0.1
pandas==2.2.2
PyPDF2==3.0.1
SQLAlchemy==2.0.32
pyodbc==5.1.0
pdfplumber==0.11.2
aiohttp==3.10.1
aiosignal==1.3.1
async-timeout==4.0.3
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-storage-blob==12.21.0
tenacity==9.0.0
tabulate==0.9.0
tabula-py==2.9.3
jpype1==1.5.0
python-dotenv